requirements.txt
streamlit
langchain
langgraph
groq
requests
grpcio
feedparser
google-api-python-client
google-genai
sentence-transformers
numpy
scikit-learn
fitz
rich
neo4j
tiktoken
fitz
beautifulsoup4
PyPDF2
matplotlib
pyvis
networkx
tavily
nltk
pinecone
python-dotenv
